Corrigindo layout da calculadora
